Remove unused standalone_map (#98)
This was code to move us from Shiny to a standalone Leaflet map. We do
still want to make this migration, but this code is now stale and we may
take a different approach when finishing the project.

Version control means that we can still refer back to the (very useful!)
code that was written, possibly even resurrecting all the code.

In the meantime, dead code makes the repository more complex to work
with, such as it not being clear what is what.