Only build detail page for changed cities (#96)
This should make the script much faster to execute.

We leverage the "Last updated" column in the database, which Google
automatically sets for us. We also check in a generated file with when
the last update happened.

This does not attempt to handle time zones, so there is a small risk of
not updating when we should. See
https://github.com/ParkingReformNetwork/mandates-map/issues/97.