Add basic JavaScript and CI infrastructure (#99)
We plan to rewrite the project from R to JavaScript, starting with the
`create_city_detail_pages.R` script. We want to migrate away from R
because contributors are less familiar with it, making it a barrier to
contributions.

We also are not happy with Shiny and want to move to a simple Leaflet
JavaScript app, like
https://github.com/ParkingReformNetwork/parking-lot-map. So, we won't
have any use for R anymore.

This PR sets up a minimal `package.json` and GitHub Actions config so
that those follow-up PRs have quality checks. It copies this config from
https://github.com/ParkingReformNetwork/parking-lot-map.

## Why JavaScript?

We know that we must use JavaScript for the Leaflet app.

I also considered using Python to rewrite our scripts, meaning that we
would have both Python and JavaScript in the repository. My motivation
is that I'm more comfortable with Python and it's ergonomic for writing
scripts. But that is not strong enough justification to support having
two languages in the repository. Instead, it's simpler for us to
consolidate around JavaScript, similar to
https://github.com/ParkingReformNetwork/parking-lot-map.