Finish implementing updateCityDetail.js (#103)
We use Handlebars for the template now. A major benefit is that all the
logic is moved into the JavaScript file. It separates presentation from
content.

This should also be substantially faster because we can parallelize all
the downloads by using promises and `Promise.all`.

The script didn't make any meaningful changes on a few sample cities,
other than minor whitespace changes.