Set up continuous deployment for city_detail/ folder (#110)
This is safe to do because we check in the generated HTML files to
version control. The only way to land a change is through a pull
request.

Unlike the Parking Lots Map, we don't use staging. That's because the
`city_detail/` folder is simple and not interactive. We have
substantially less risk that any PR breaks those pages.