Migrate from jest to playwright (#147)
Resolves #140 

### Added ###
- playwright dependency 
- playwright config
- similar to Parking Lots map, but with setting NODE_ENV and increased
timeout for web server
- tests folder with playwright tests

### Removed ###
- jest, babel, babel-jest dependencies

### Testing ###
`npm test` runs like before. However, I noticed it's taking quite a long
time for the server to start (had to increase timeout threshold). I
suspect it's bundling all the .png files? Not sure, but I think it would
be something we optimize in the future.

---------

Co-authored-by: tunglinn <tunglin@yahoo.com>