Revert bad CSV updates (#174)
Part of fixing
https://github.com/ParkingReformNetwork/mandates-map/issues/168. As
explained there, the Google Table got corrupted so both these PRs
resulted in an empty CSV.

We need to revert the changes entirely because we must go back to the
last valid state of the CSVs. This is because our CSVs store the
latitude and longitude in them, which we don't have in the Google Table.
We only grab the lat/long for brand new cities. So we need to get back
the historical entries.

This revert results in `scripts/city_detail_last_updated.txt` going back
to July. So the next GitHub Action run will pull in the past few weeks
of changes.