Rename map from "Mandates Map" to "Reform Map" (#180)
We decided that this is a much more accurate description of what the map
is about: where reforms have happened. Compared to
https://github.com/ParkingReformNetwork/parking-requirement-database
being about where mandates still exist.

Reform is also a more positive framing. We want to show how much
progress is being made.

We don't rename the R map because we're sprinting to drop it in favor of
JavaScript.