Add searchbox (#199)
Closes https://github.com/ParkingReformNetwork/reform-map/issues/185.

We use https://selectize.dev/, as Shiny does under the hood.

Note that this uses a dedicated HTML element with `position: absolute`,
same as Shiny, rather than using a Leaflet control like we did for the
Legend. I tried to get the Leaflet control to work but there were issues
like scrolling not working and interacting with the dropdown resulting
in dragging the map. The map has too much going on with event
propagation. The main benefit of using a Control would have been simpler
positioning, but our SCSS code works fine.

Search is not finished:

* https://github.com/ParkingReformNetwork/reform-map/issues/195
* https://github.com/ParkingReformNetwork/reform-map/issues/198