Connect search and filter functionality (#235)
Before, each filter was being traeted as its own separate thing, without
paying attention to the other filters and the search. They would each
call `changeSelectedMarkers` with their own `filterFn`. So if you
changed one filter, it would ignore the other filters and override their
prior filtering.

Instead, when deciding whether to show a city, we need to consider _all_
the filters. The city must match every filter.

## Search vs filter

I wasn't sure how we should handle search vs. the filter mechanism. If
you have selected a city in the search, but it doesn't match the
filters, what should we do?

ChatGPT helped convince me that search should take priority. If you
actively choose a city, then the least confusing UX is to still show
that city, regardless of filters.

## Punts on population slider

I was having a tough time figuring out how to hook up the code to the
filter. I want to refactor the code first. So this PR punts on that
file.