Deduplicate sliders code in population slider (#244)
Before, we searched for the slider HTML elements several times with
query selectors. Now, we only find it once and store the three elements
in a `PopulationSliders` object.

`PopulationSliders ` is set up as a
[class](https://www.typescriptlang.org/docs/handbook/2/classes.html)
with the method `getCurrentIndexes(): [number, number]`. That will be
helpful in
https://github.com/ParkingReformNetwork/reform-map/issues/236, where we
need to get the current indexes inside the function `shouldBeRendered`
from `filter.ts`. I didn't want to leak that logic into `filter.ts`, so
using a method allows us to "encapsulate" it instead in the method.

This also inlines the `init` function directly into
`setUpPopulationSlider`. It doesn't change the code, other than to use
the `PopulationSliders ` object.