Add no requirements toggle (#286)
Closes https://github.com/ParkingReformNetwork/reform-map/issues/187.
Turns out this feature was really important.

This toggle is simply another filter, but it appears on its own rather
than being in the filter popup. That's a little weird, but we need to
keep it for now until we can redesign the UI.

Also fixes the border color for the filter icon to match Leaflet's gray
color.

<img width="363" alt="Captura de pantalla 2023-10-15 a la(s) 10 02 14 p
m"
src="https://github.com/ParkingReformNetwork/reform-map/assets/14852634/8df47fb6-d310-4ddc-93e4-b0aa5ad05ba3">

The UI is a little awkward because search will cover the toggle, just
like the old UI. Tony said that's fine for now.

Also, it's only a checkbox and not yet a toggle to unblock adding this
critical feature. We need to switch to a toggle in
https://github.com/ParkingReformNetwork/reform-map/issues/285.