Fix missing HTML for cities without state (#352)
For countries like Canada, we don't set the state since it wouldn't make
sense. We need to create HTML entries for them though. I manually
invoked the script to run on those cities.

This also manually fixes La Rochelle FR and London, which had their
states change.

Finally, this adds a new test to ensure that every `citation_url`
actually exists. But it's skipped for now because it fails.