Remove share icon (#392)
Unlike the Parking Lot Map, this icon is not very useful. It always
copies the same URL: whatever is already in the URL bar. The typical
user will know how to copy the URL from the browser. This is unlike PLM,
where the share icon saves the selected city.

Removing this icon frees up valuable screen real estate, especially on
mobile. I'm not sure yet what we will do with this screen real estate,
but it will be useful with the revamp. In the worst case, it is
whitespace, which is very valuable for a clean UI.

<img width="378" alt="Screenshot 2024-07-17 at 8 21 38 PM"
src="https://github.com/user-attachments/assets/ef6e279d-3bc3-4863-8903-b36f972c657c">