Filter updates the table (#434)
This sets up all the state management for the table. We load all the
data on initial setup, then use Tabulator's filter mechanism to update
what is rendered based on `FilterState`.

We still need to figure out which columsn to render, along with a bunch
of other improvements. But the state management is now complete!
(Outside of maybe a counter.)

This also enables sorting.