Set up remaining columns for table (#439)
Some decisions:

* Don't show the description. It would make it impossible to fit
everything in one line, which gets in the way of scanning the table
* Call it "Place" rather than "City" to account for
states/countries/counties. This will need to be improved throughout the
data model in general, but this is a good first step. Note that the
Place is always defined in our dataset currently.
* Freeze the Place column because this is useful on mobile
* Don't hide any rows on mobile - it may still be useful to see them
* Let the users resize and rearrange columns
* Use fixed width, and don't show the entire row by default. The edge
cases make the table too sparse otherwise if we want to fit everything.
Users can either resize the column or go to the details page

<img width="1374" alt="Screenshot 2024-08-03 at 7 56 34 PM"
src="https://github.com/user-attachments/assets/14fe5ce5-8bd3-4e47-85e8-6c190253d380">

<img width="376" alt="Screenshot 2024-08-03 at 7 56 48 PM"
src="https://github.com/user-attachments/assets/ea8cdeb3-dca2-405d-ac86-3959c8ce35a8">

The app is too slow when there are lots of rows. It was slow before, and
this PR makes it even worse where it hangs. This needs to be fixed
before the table view goes live.