Move population above policy reform type (#445)
Now that we select all the reform types by default, it's less important
that we show it right away.

The benefit of this change is that all checkboxes now appear together,
which makes more sense visually. It's also prework for the filter
redesign, where I'm proposing putting the checkboxes behind expandable
accordions.