Fix bug with population slider having bad size (#449)
Before, you could get into a weird state with the population slider
where it didn't have proper width. That's because the slider requires
the popup is open to calculate widths. So, the slider needs to be
redrawn every time the filter popup is opened up, since the values might
have changed since the last time the popup was one.