Generate filter checkbox options (#451)
This is prework for converting the checkbox options into accordions.
That code is quite complex, so we'll want to generate it with JavaScript
rather than hardcoding HTML.

This also DRYs setting which options are pre-selected vs. what is the
universe of valid options. We no longer duplicate the option names in
main.ts.