Parse array fields on initial load (#458)
CSVs can't easily represent `string[]`, so we've been using `, ` as the
delimiter between values. That's been requiring us to call `.split(',
')` at runtime, which is less readable and also a waste of performance
to keep doing that.

I considered migrating to JSON, but it's important we still have the CSV
for researchers who need the data set. Maintaining both JSON and CSV
would be annoying.

Instead, we now normalize/parse the data on the initial app load. This
model will be helpful for future projects like filtering by year because
we can parse the date time to Luxon dates on the initial load.

The table keeps the same sorting semantics as before.