Revamp population slider (#476)
The touch targets have been way too small for the population slider. But
the complication was that the legend is too crowded to make anything
bigger. The internet shows lots of designs without the legend, and it
made me realize it's not necessary.

We now also have more flexibility to add more ranges if we want.

Before:


![image](https://github.com/user-attachments/assets/59f6b770-1641-4da2-a876-49691a1b74ba)

After:


![image](https://github.com/user-attachments/assets/6dfb155e-3427-4507-988d-7de2a96245ad)