Use full names for countries (#478)
This is much more accessible.

It's plausible we will still want to use the two-letter code for the
citation_url once revisiting how those URLs are formed. We can still do
that with this implementation because the database still uses the two
letter code; in `data.ts`, we can save for example `country_code` as
another field.