Always generate HTML pages (#496)
Should close
https://github.com/ParkingReformNetwork/reform-map/issues/354.

It's very fast to generate the HTML pages. There was no good reason to
skip that generation every time we run the script. The only thing that
is slow is downloading citation images. To know whether to skip that, we
only need to look at the "last updated" value for citations, not for the
report and city.

Independent of
https://github.com/ParkingReformNetwork/reform-map/issues/354, this
change is necessary to unblock us generating the HTML pages through CI
rather than at PR-time.