Remove `updated` from extended.json (#498)
We don't actually use `updated` in the CSV or HTML pages, so there is no
need to save it in `extended.json`. We can add it back if the volunteer
team wants it, but for now I want to simplify things as much as
possible.

This also fixes the type hints for `ExtendedEntry`.