Store enums as lowercase (#530)
Directus is going to store everything as lowercase, which is a more
natural way to describe the data. For example, in Table view, it makes
sense to be lowercase.

The UI is updated for the filter options to still show as Sentence case,
but the underlying data & table view use lowercase.

This PR also renames some `requirements` like `ADU` and `TDM`.