Normalize file extensions (#537)
We now enforce that files need to be .pdf, .jpg, .png, or .docx. We
lowercase JPG and PNG, and fix .jpg_medium to be .jgp. In the database,
I converted a .eml file to be a .pdf. This will help with the Directus
migration.

Likewise, I deleted all files and regenerated them to ensure we have the
latest copies. I added a test to make sure that every expected file is
actually present.