Write missing coordinates to Directus (#541)
We will continue to find the coordinates automatically through reverse
geocoding. It's not easy to set that up in Directus, so we'll write to
Directus any missing coordinates during the nightly sync.

This also sets up the basic skeleton for `npm run sync-directus` by
reading the Places table. Directus operates in terms of a numeric ID for
places, but our JSON files expect a string `PlaceId`. So this new
function returns a helpful mapping between those two key types.