App can handle multiple policies per place (#566)
With this change, it's now safe for the `data/json` files to not set
`legacy` and to have multiple of `reduce_min`, `rm_min`, and `add_max`.

The key constraint is that they must set `legacy`; otherwise, they must
set exactly one of the new-style policies. This allows our scripts and
web app to use the `unifiedPolicy` field, which emulates the legacy
style of reform.

A follow up will update the database and `sync-directus` to actually
start populating the JSON with these new policy records. This PR only
prepares everything else so the JSON change won't break it all.