Allow getting the processed new style policy records (#579)
It's now possible to have a `ProcessedCoreEntry` and
`ProcessedCompleteEntry` that includes both `unifiedPolicy` and the new
style policy records. This will allow the app and scripts to choose
which style of data they want to work on.

We use a feature flag on `processRawCoreEntry` to avoid slowing down
production by reading in unused data. The scripts read the full data set
because performance doesn't matter there.