FilterState supports multiple policy records (#583)
The backend of the app now supports the main change for having multiple
policy records per place. There's a new type `PolicyTypeFilter` in the
`FilterState` for what policy type you're focusing on, between the three
reform types, "any parking reform", and "legacy reform". The app is
hardcoded to use "legacy reform", but we can proceed with implementing
the four other types meanwhile behind-the-scenes. Users also have no way
to change the policy type yet.

PlaceFilterManager now has to keep track of which specific records are
matched, not only whether a place is matched. We do that with this new
type:

```typescript
export interface MatchedPolicyRecords {
  rmMinIdx: number[];
  reduceMinIdx: number[];
  addMaxIdx: number[];
}
```

This allows us to decide which records to show based on the filter
efficiently. Using indexes avoids duplicating the information in memory.