In counter, only mention policy records when relevant (#586)
Tony and I realized it was confusing and noisy to talk about # of policy
records on the map view since the map view only shows places as a single
dot. It makes no difference to the map if there are >1 records per
place.

The # of policy records is relevant to table view, but only if the # of
policy records > # of places. In that case, we still show the full text
but with a longer description explaining what is happening.