Change test strategy for HTML pages (#644)
We're migrating to Eleventy for HTML page generation rather than the
custom script. Eleventy doesn't have a good way to generate based on
test data, so we need to switch to testing actual pages.

Now that we test on real data, we need to use `npm run gen-html` before
`npm test`. I considered splitting up a different `test-gen-html`
command, but I think it's too complex and not worth it given that HTML
page generation is really fast.