Extract CSS for details to dedicated file (#647)
Uses this technique: https://www.11ty.dev/docs/quicktips/inline-css/.
This allows us to have 1+ normal CSS files outside of the HTML page,
while still inlining it in the final result. The inlining is useful for
performance and to avoid any caching issues that the CSS file was cached
incorrectly by the browser, given that we don't use a build tool like
Parcel for the details pages.

The CSS file gets minified by `clean-css`, so we don't need to worry
about using whitespace and comments in the source code.