Hide proposed and repealed reforms in 'any parking reforms' (#696)
As explained in
https://github.com/ParkingReformNetwork/reform-map/pull/686, our
handling of the Reform Status is off. Eventually, we plan to have a
dropdown to choose which data set to load: passed, proposed, or
repealed.

In the meantime, the views for "add max", "rm min", and "reduce min" do
have a Reform Status option that lets you union the three statuses.
However, we cannot sensibly expose a status option with 'any parking
reforms' because the metadata is policy-record-specific rather than
place-specific. So, this PR instead completely removes places with only
proposed/repealed reforms from 'any parking reforms'. It removes them
from the table view (universe of valid places) & also FilterState so
they don't match in the map either.