Improve counter text for table view (#706)
With the revamp, the table no longer shows general information about
places. Instead, it shows information about the reforms themselves. So,
we should clarify with the counter what you're seeing.

This also clarifies what you're seeing with search when on the table.

Per https://github.com/ParkingReformNetwork/reform-map/pull/696, when
"any parking reform" is chosen, we nuance that the data set is only for
"adopted" reforms.