Don't hide scope & land use filters with parking maximums (#708)
Fixes https://github.com/ParkingReformNetwork/reform-map/pull/696. 

It makes sense to hide "Reform scope" and "Affected land use" when the
policy type is minimum removals and "all minimums repealed" is checked.
However, we do need to fix FilterState to ignore those two fields in
this edge case correctly.

We should not hide the two option groups with "add parking maximums",
however, because the options are checking the maximum policy records,
not the minimum removal records.