Set default policy type to remove minimums (#734)
Both Etienne and Tony independently made this request because they want
the table view to show date of reform, which is not possible if the
default policy type is "any parking reform".

They understand this changes the experience when you uncheck "only
places with all parking minimums removed". They are okay with this.