Determine option values for each data set (#738)
In a follow-up PR, I want to improve our UX so that the filter only
shows options actually relevant to the current dataset. For example, we
shouldn't show years that aren't even in the dataset.

This PR refactors our code to determine the option values. We don't yet
use the new sets.