Refactor default options behavior (#739)
Generally, with the filter, every option should be chosen by default.
(We will add an ability to save/load a pre-selected view.)

The only exception is `status` because it doesn't make sense to have >1
status. That's the only option we were doing something unusual with.
However, `status` will be reworked entirely to be a dropdown rather than
checkboxes.

This PR is meant to make the modeling a little more clear and reduce
complexity in preparation for hiding filter options based on the
dataset.