Fix issues with structuredClone in tests (#741)
Follow up to
https://github.com/ParkingReformNetwork/reform-map/pull/707. Turns out
`structuredClone` breaks with our custom `Date` class because it tries
to convert `Date.parsed` (Luxon DateTime) into an object.

So, now we use factory functions that create a new object for the state
and entries each time. This is a little less efficient, but it's a test
so it's fine and better to avoid the confusion of accidentally sharing
state between tests.