Filter options only show values relevant to the current data set (#763)
For example, the country list will only show countries in the data set,
like "any parking reform" vs "reduce parking minimums".

This was a general improvement I wanted to make, but it will be
especially important when adding parking benefit districts.

The core idea of this PR is to mark checkboxes as `hidden` if they are
not in the dataset. The checkbox still exists and its prior check state
is still preserved in `FilterState`, only the user cannot interact with
it. Then, we update the logic for these places to only operate on
visible checkboxes:

- supplemental title, e.g. `(4/5)'
- 'check all' button
- 'uncheck all' button

## No caching

The original implementation attempted to use a caching mechanism to
avoid performance concerns. `FilterOptions` would store the prior data
set ID and only run the code for the visibility of checkboxes when the
dataset changed. However, I realized this solution didn't work properly
because the logic for visibility lives in each distinct filter group,
rather than the overall filter, and there was a race condition that one
filter group would try to update the global `FilterOptions` while other
filter groups were still running. We would need some type of lock to
coordinate changing the cache between all the filter groups. This
particular code path didn't even show up in benchmarking, so I decided
the cache wasn't worth it.