Refactor how merged options in the filter are set up (#768)
Prework for making `status` a dropdown rather than checkbox. That will
result in 12 datasets (4 policy type x 3 status). This change preps the
filter options code, i.e. the code for the interactive filter component.

We won't be able to use `any parking reform` as a standin for `merged`
(the whole universe), so this sets up a distinct type.

This also simplifies the class `FilterOptions` to use a new `datasets`
value, rather than a separate value per dataset. It's much more concise.