Make app resilient to PlaceId having a country (#796)
Prep for https://github.com/ParkingReformNetwork/reform-map/issues/484.
This fixes two places that would break when we add the country to the
place ID:

* Scorecards and markers. We want to avoid the country being in the
marker's tooltip because it's overly verbose. But the scorecard wad
epending on the tooltip to determine what place the marker corresponds
to. This is now decoupled.
* URL for the details page. We shouldn't change the URL to incorporate
the country until we have redirects set up and the change is fully
ready.