Search splits out place, state, and country (#801)
Part of https://github.com/ParkingReformNetwork/reform-map/issues/484.
We want the country to be included in search. But we can improve
Choices.js's algorithm by treating the place, state, and country as
distinct strings so that you can do something like type "Germany" and
get all places in Germany.

This doesn't yet change the value we show to the user, which will be a
follow up.