Store the URL encoding in core.json (#805)
This unblocks
https://github.com/ParkingReformNetwork/reform-map/issues/804. It will
allow us to start using the new scheme for brand new places.

We will also be able to update the database names for existing places
while preserving the prior URL, so we do not need a redirect. To do
this, a follow-up PR will preserve the prior encoding when calling
sync-directus.