Use any parking reform as default (#794)
Now that we have "views"—aka a way to load the map from query
parameters—we no longer need to special-case the initial view as all
places with total repeals. We can use the more sensible default of any
parking reform, and use a view saved to the repeals for embedding in an
iframe.